% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_groups.R
\name{create_groups}
\alias{create_groups}
\title{Create groups}
\usage{
create_groups(data, n_groups, select_contains)
}
\arguments{
\item{data}{}

\item{n_groups}{}

\item{select_contains}{}
}
\value{

}
\description{
Create groups
}
