% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_to_equity.R
\name{debt_to_equity}
\alias{debt_to_equity}
\title{Debt-to-equity}
\usage{
debt_to_equity(
  url = "http://www.stern.nyu.edu/~adamodar/pc/datasets/dbtfund.xls",
  industry
)
}
\arguments{
\item{url}{}

\item{industry}{}
}
\value{

}
\description{
Debt-to-equity
}
