% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fcff.R
\name{estimate_fcff}
\alias{estimate_fcff}
\title{Estimate FCFF}
\usage{
estimate_fcff(
  data,
  revenue_growth,
  ppe_rate,
  depreciation_rate,
  net_operating_profit_margin,
  nwc_to_revenues,
  explicit_forecast_period
)
}
\arguments{
\item{data}{}

\item{revenue_growth}{}

\item{ppe_rate}{}

\item{depreciation_rate}{}

\item{net_operating_profit_margin}{}

\item{nwc_to_revenues}{}

\item{explicit_forecast_period}{}
}
\value{

}
\description{
Estimate FCFF
}
