% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_market_value.R
\name{estimate_market_value}
\alias{estimate_market_value}
\title{Estimate market value}
\usage{
estimate_market_value(
  data,
  base_year,
  explicit_forecast_period,
  growth_perpetuity
)
}
\arguments{
\item{data}{}

\item{explicit_forecast_period}{}

\item{growth_perpetuity}{}
}
\value{

}
\description{
Estimate market value
}
