% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot growth}
\usage{
plot_growth(data, no_of_years = 3)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
TODO: fiks label, legg inn rolling mean (3 eller 5 år) på revenue_growth og nopm
}
