% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_value_creation.R
\name{plot_value_creation}
\alias{plot_value_creation}
\title{Plot value creation}
\usage{
plot_value_creation(data, avkastning = "roic", avkastningskrav = "wacc")
}
\arguments{
\item{data}{}

\item{avkastningskrav}{}
}
\value{

}
\description{
Plot value creation
}
\details{
ROE > ROIC --> belåner operasjonelle eiendeler som gir avkastningfinansieringskostnaden (kd)
}
