% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prep data}
\usage{
prep_data(
  data,
  balance_sheet = FALSE,
  columns_to_remove = c("Consolidatet statement", "Start date", "End date", "0",
    "Currency")
)
}
\arguments{
\item{data}{}

\item{columns_to_remove}{}
}
\value{

}
\description{
Prep data
}
