% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_free_rate.R
\name{risk_free_rate}
\alias{risk_free_rate}
\title{Risk-free rate}
\usage{
risk_free_rate(country = "Norway")
}
\arguments{
\item{country}{}
}
\value{

}
\description{
Returns the risk-free rate, define as the 10-year government bond rate, based
on country and year.
}
\details{
TODO: remove warning messages and consider changing to full dataset (faster) in stead of running function on each row
}
