% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlevered_beta.R
\name{unlevered_beta}
\alias{unlevered_beta}
\title{Unlevered beta}
\usage{
unlevered_beta(
  url = "http://www.stern.nyu.edu/~adamodar/pc/datasets/betas.xls",
  industry,
  time_period = "last",
  ...
)
}
\arguments{
\item{url}{}

\item{industry}{}

\item{time_period}{}

\item{...}{}
}
\value{

}
\description{
Unlevered beta
}
