% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_capital.R
\name{wacc}
\alias{wacc}
\title{WACC}
\usage{
wacc(debt_to_equity, cost_of_debt, cost_of_equity, tax_rate)
}
\arguments{
\item{debt_to_equity}{}

\item{cost_of_debt}{}

\item{cost_of_equity}{}

\item{tax_rate}{}
}
\value{

}
\description{
Calculates the weighted average cost of capital
}
