% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_capital.R
\name{cost_of_equity}
\alias{cost_of_equity}
\title{Cost of equity}
\usage{
cost_of_equity(rf, beta, market_risk_premium)
}
\arguments{
\item{rf}{}

\item{beta}{}

\item{market_risk_premium}{}
}
\value{

}
\description{
Cost of equity
}
